refactor(observability): metrics.ts を関心ごとに分割する
404行の metrics.ts（max-classes-per-file/max-lines 抑制）を7ファイルに分割し、
metrics.ts は barrel（再エクスポート）として残す。

- metric-names.ts (METRIC) / prometheus-labels.ts (labelsToKey) /
  token-metrics.ts (recordTokenMetrics) / agent-labels.ts (infer*/buildAgentMetricLabels) /
  error-classification.ts (classifyErrorType) / prometheus-collector.ts /
  prometheus-server.ts。
- @vicissitude/observability/metrics の公開シンボル集合を完全保存（barrel）。
  26 consumer は無変更。新サブパス export は追加せず（YAGNI）。
- max-classes-per-file/max-lines 抑制を解消。

logger.ts の pino any 縮減は別関心のため別 Issue。
nr validate / nr test (2651 pass, 0 fail) green。
